requirements.txt
pandas==2.2.3
pyyaml==6.0
numpy==1.26.0
pytest==8.2.0
python-dotenv==1.0.0
